feat(frontend): Now redirect to clone vs vv file
The above proposal may cause inconvenience. It may happen that the client's browser does not automatically open the .vv file,
Now redirect to the homepage and start the VM automatically.

This operation apply for all users except admin.

TO-DO: read the variable auto_view = 1 to open the .vv file and avoid user interaction by clicking on the button.

Another possibility is to enable the auto_view variable in rvd_front.conf but this would apply to users with more than one machine.
The issue requests should only be applied to a single machine that starts directly.

Issue #1145